% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corum.R
\name{corum2graphlist}
\alias{corum2graphlist}
\title{basic data structure for sets of complexes or sets of pull-downs:
a list of graph instances}
\usage{
corum2graphlist(
  corum.df,
  organism = "Human",
  subunit.id.type = c("UNIPROT", "ENTREZID")
)
}
\description{
basic data structure for sets of complexes or sets of pull-downs:
a list of graph instances
}
