% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome.R
\name{getBioplexProteome}
\alias{getBioplexProteome}
\title{Obtain BioPlex3 proteome data}
\usage{
getBioplexProteome(cache = TRUE)
}
\arguments{
\item{cache}{logical. Should a locally cached version used if available?
Defaults to \code{TRUE}.}
}
\value{
A \code{\linkS4class{SummarizedExperiment}} storing protein expression
data for the both cell line(s) with 5 replicates each.
}
\description{
Functionality for retrieving the BioPlex3 protein expression
data comparing expression in the HCT116 and the 293T cell lines.
}
\examples{

  se <- getBioplexProteome()
  
}
\references{
BioPlex: \url{https://bioplex.hms.harvard.edu}
}
