% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{getGSE122425}
\alias{getGSE122425}
\title{Convenient access to 293T transcriptome data from GEO}
\usage{
getGSE122425(cache = TRUE)
}
\arguments{
\item{cache}{logical. Should a locally cached version used if available?
Defaults to \code{TRUE}.}
}
\value{
A \code{\linkS4class{SummarizedExperiment}} storing RNA-seq
data for the 293T cell line.
}
\description{
Functionality for storing the 293T RNA-seq data
from GSE122425 in a
\code{\linkS4class{SummarizedExperiment}}. The dataset includes three
wild type samples and three NSUN2 knockout samples.
}
\examples{

  # Obtain the data as a SummarizedExperiment
  se <- getGSE122425()
 
}
\references{
GSE122425: \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE122425}
}
