% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corum.R
\name{getCorum}
\alias{getCorum}
\title{Obtain CORUM protein complex data}
\usage{
getCorum(
  set = c("all", "core", "splice"),
  organism = "Human",
  remap.gene.ids = FALSE,
  cache = TRUE
)
}
\arguments{
\item{set}{character. Valid options include:
\itemize{
\item \code{"all"}: complete set of complexes,
\item \code{"core"}: core set of complexes,
\item \code{"splice"}: complexes with splice variants.
}
Defaults to \code{"all"}.}

\item{cache}{logical. Should a locally cached version used if available?
Defaults to \code{TRUE}.}
}
\value{
A \code{data.frame}.
}
\description{
Functionality for retrieving the CORUM protein complex data.
Available complex collections include:
\itemize{
\item complete set of complexes,
\item core set of complexes,
\item complexes with splice variants.
} See references.
}
\examples{
# Obtain the core set of CORUM complexes
core <- getCorum(set = "core")
}
\references{
CORUM: \url{http://mips.helmholtz-muenchen.de/corum/#download}
}
